Extracted methods for collecting metadata and handling the editing script for better maintenance. Also added new methods for extracting metadata
